Implemented OutputStream
The OutputStream can be piped to by the SegmentStream. It is associated with a Writer implementation, which takes the segments and writes them out to the desired target. The first writer will be the HttpServerWriter. There are no limitations to this. Therefore other writer targets may be created in the future